+++ core/lib/config/src/configs/mod.rs
@@ -6,7 +6,7 @@ pub use self::{
     contract_verifier::ContractVerifierConfig,
     contracts::chain::ContractsConfig,
     da_client::{
-        avail::AvailConfig, celestia::CelestiaConfig, eigenda::EigenConfig, DAClientConfig,
+        avail::AvailConfig, celestia::CelestiaConfig, eigen::EigenConfig, DAClientConfig,
     },
     da_dispatcher::DADispatcherConfig,
     database::{DBConfig, PostgresConfig},
