+++ core/node/da_clients/src/lib.rs
@@ -1,6 +1,6 @@
 pub mod avail;
 pub mod celestia;
-pub mod eigen_da;
+pub mod eigen;
 pub mod no_da;
 pub mod node;
 pub mod object_store;
