+++ core/node/da_clients/src/node/mod.rs
@@ -1,10 +1,10 @@
 pub use self::{
-    avail::AvailWiringLayer, celestia::CelestiaWiringLayer, eigenda::EigenWiringLayer,
+    avail::AvailWiringLayer, celestia::CelestiaWiringLayer, eigen::EigenWiringLayer,
     no_da::NoDAClientWiringLayer, object_store::ObjectStorageClientWiringLayer,
 };
 
 mod avail;
 mod celestia;
-mod eigenda;
+mod eigen;
 mod no_da;
 mod object_store;
